BACK/backend/requirements.txt
asgiref==3.8.1
astroid==3.2.4
colorama==0.4.6
dill==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
isort==5.13.2
joblib==1.4.2
mccabe==0.7.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tomlkit==0.13.2
tzdata==2024.1
